Prefers exact match to support annotationQuery; Makes it work with C* 3.0.8
The annotationQuery in the http api is an exact match; there's no
full-text search feature, and we also don't document our encoding rules
such that another party could use such a feature. This reverts to exact
match.

This also removes the SASI index on duration, which allows the model to be
usable on a supported version of Cassandra, either 3.0.8 or 3.8+